% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_add_node.R
\name{map_add_node}
\alias{map_add_node}
\title{Add a node to the mental-health map}
\usage{
map_add_node(current_matrix, new_factor)
}
\arguments{
\item{current_matrix}{An unweighted directed matrix.}

\item{new_factor}{String input of the new factor the user wants to add.}
}
\value{
An updated unweighted directed matrix including the new factor.
}
\description{
Adds a node the user has selected to the existing map.
}
\examples{
current_matrix <- matrix(0, nrow = 3, ncol = 3)
colnames(current_matrix) <- rownames(current_matrix) <- c("Stress", "Worrying", "Breakup")
new_factor <- "Tired"
map_add_node(current_matrix, new_factor)
}
